// src/types/editorVars.ts
export enum BootType {
  Box = "box",
  Recovery = "recovery",
}

export enum ModeOptions {
  //doors
  Aperture = "aperture" ,
  Simple = "simple",
  Double = "double",
  Pocket = "pocket",
  //windows
  Fix = "fix",
  Flap = "flap",
  Twin = "twin",
  Bay = "bay",

  // Stair
  SimpleStair = "simpleStair",

  // Office furniture

  OfficeTable = "officeTable",
}

export enum BinderType {
  Segment = "segment",
  Node = "node",
  Obj = "obj",
  BoundingBox = "boundingBox",
}